+++ tests/testthat/test-utils-checks.R
@@ -1,24 +1,24 @@
-testthat::test_that("`test_nested()` | General test", {
+testthat::test_that("`test_nested_list()` | General test", {
   list(1, 1) |>
-    test_nested() |>
+    test_nested_list() |>
     testthat::expect_false()
 
   list(1, 1:2) |>
-    test_nested() |>
+    test_nested_list() |>
     testthat::expect_true()
 
   "a" |>
-    test_nested() |>
+    test_nested_list() |>
     testthat::expect_false()
 })
 
-testthat::test_that("`test_one_depth()` | General test", {
+testthat::test_that("`test_one_depth_list()` | General test", {
   list(1, 2, 3) |>
-    test_one_depth() |>
+    test_one_depth_list() |>
     testthat::expect_true()
 
   list(list(1, 2, 3)) |>
-    test_one_depth() |>
+    test_one_depth_list() |>
     testthat::expect_false()
 })